+++ tests/testapp/translated_articles_urls.py
@@ -1,5 +1,5 @@
-from django.conf.urls import url
 from django.shortcuts import get_object_or_404
+from django.urls import path
 
 from feincms3.applications import page_for_app_request
 from feincms3.shortcuts import render_detail
@@ -16,4 +16,4 @@ def detail(request, pk):
 
 
 app_name = "translated-articles"
-urlpatterns = [url(r"^(?P<pk>[0-9]+)/$", detail, name="detail")]
+urlpatterns = [path("<int:pk>/", detail, name="detail")]